+++ Source/MCLib/MLR/MLR_I_L_DeT_PMesh.cpp
@@ -152,9 +152,9 @@ void
 	Check_Object(this); 
 	Check_Pointer(data);
 
-	Verify(coords.GetLength() == 0 || dataSize == coords.GetLength());
-	Verify(colors.GetLength() == 0 || dataSize == colors.GetLength());
-	//Verify(texCoords.GetLength() == 0 || dataSize == texCoords.GetLength());
+	Verify(coords.GetLength() == 0 || dataSize == (int)coords.GetLength());
+	Verify(colors.GetLength() == 0 || dataSize == (int)colors.GetLength());
+	//Verify(texCoords.GetLength() == 0 || dataSize == (int)texCoords.GetLength());
 
 	normals.AssignData(data, dataSize);
 }
@@ -188,8 +188,8 @@ void
 	Check_Pointer(data);
 	Verify(gos_GetCurrentHeap() == Heap);
 
-	Verify(coords.GetLength() == 0 || dataSize == coords.GetLength());
-	Verify(texCoords.GetLength() == 0 || dataSize == texCoords.GetLength() || dataSize == 2*texCoords.GetLength() );
+	Verify(coords.GetLength() == 0 || dataSize == (int)coords.GetLength());
+	Verify(texCoords.GetLength() == 0 || dataSize == (int)texCoords.GetLength() || dataSize == 2 * (int)texCoords.GetLength());
 
 	litColors.SetLength(dataSize);
 	colors.AssignData(data, dataSize);
